+++ src/main/java/com/pricing/ordering/InNDaysTime.java
@@ -1,4 +1,4 @@
-package com.pricing.domain;
+package com.pricing.ordering;
 
 import lombok.Value;
