+++ src/main/java/com/pricing/ordering/Today.java
@@ -1,4 +1,4 @@
-package com.pricing.domain;
+package com.pricing.ordering;
 
 import lombok.Value;
