+++ src/main/java/com/pricing/ordering/When.java
@@ -1,4 +1,4 @@
-package com.pricing.domain;
+package com.pricing.ordering;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
