+++ src/main/java/com/pricing/product/ProductUnit.java
@@ -1,4 +1,4 @@
-package com.pricing.domain;
+package com.pricing.product;
 
 import lombok.Getter;
