+++ src/main/java/com/pricing/stock/Apples.java
@@ -1,9 +1,11 @@
-package com.pricing.domain;
+package com.pricing.stock;
 
+import com.pricing.product.Pence;
+import com.pricing.product.Product;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
 
-import static com.pricing.domain.ProductUnit.Single;
+import static com.pricing.product.ProductUnit.Single;
 
 @Value
 @EqualsAndHashCode(callSuper = true)
