+++ src/main/java/com/pricing/stock/Bread.java
@@ -1,9 +1,11 @@
-package com.pricing.domain;
+package com.pricing.stock;
 
+import com.pricing.product.Pence;
+import com.pricing.product.Product;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
 
-import static com.pricing.domain.ProductUnit.Loaf;
+import static com.pricing.product.ProductUnit.Loaf;
 
 @Value
 @EqualsAndHashCode(callSuper = true)
