+++ src/main/java/com/pricing/stock/Milk.java
@@ -1,10 +1,11 @@
-package com.pricing.domain;
+package com.pricing.stock;
 
+import com.pricing.product.Pence;
+import com.pricing.product.Product;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
 
-import static com.pricing.domain.ProductUnit.Bottle;
-import static com.pricing.domain.ProductUnit.Loaf;
+import static com.pricing.product.ProductUnit.Bottle;
 
 @Value
 @EqualsAndHashCode(callSuper = true)
