+++ src/test/java/com/pricing/stock/ApplesStockItemTests.java
@@ -1,10 +1,10 @@
-package com.pricing.domain;
+package com.pricing.stock;
 
 import org.junit.jupiter.api.Test;
 
-import static com.pricing.domain.Apples.apples;
-import static com.pricing.domain.Pence.pence;
-import static com.pricing.domain.ProductUnit.Single;
+import static com.pricing.stock.Apples.apples;
+import static com.pricing.product.Pence.pence;
+import static com.pricing.product.ProductUnit.Single;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class ApplesStockItemTests {
