+++ src/test/java/com/pricing/stock/MilkStockItemTests.java
@@ -1,10 +1,10 @@
-package com.pricing.domain;
+package com.pricing.stock;
 
 import org.junit.jupiter.api.Test;
 
-import static com.pricing.domain.Milk.milk;
-import static com.pricing.domain.Pence.pence;
-import static com.pricing.domain.ProductUnit.Bottle;
+import static com.pricing.stock.Milk.milk;
+import static com.pricing.product.Pence.pence;
+import static com.pricing.product.ProductUnit.Bottle;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class MilkStockItemTests {
